+++ monkey/common/agent_configuration/default_agent_configuration.py
@@ -8,6 +8,7 @@ from .agent_sub_configurations import (
     ICMPScanConfiguration,
     NetworkScanConfiguration,
     PluginConfiguration,
+    PolymorphismConfiguration,
     PropagationConfiguration,
     ScanTargetConfiguration,
     TCPScanConfiguration,
@@ -92,6 +93,7 @@ DEFAULT_AGENT_CONFIGURATION = AgentConfiguration(
     credentials_collectors=CREDENTIALS_COLLECTORS,
     payloads=PAYLOAD_CONFIGURATION,
     propagation=PROPAGATION_CONFIGURATION,
+    polymorphism=PolymorphismConfiguration(randomized_agent_hash=False),
 )
 
 DEFAULT_RANSOMWARE_AGENT_CONFIGURATION = deepcopy(DEFAULT_AGENT_CONFIGURATION)